[REF] website_sale_category_megamenu: Use customize menu
Split functionalities in different templates that can be enabled or
disabled in the Customize menu in the website.

Create a new division in the html where mega menu and other
functionalities are placed.

Facilitate the extension of this module by having specific templates to
insert functionalities in the parent view. So that new component can be
moved easier by another module.

Replace the usage of `@class` by `hasclass`.

This module is like this more generic.