[ADD] cet_website_sale: Stock in e-commerce
Prevent ordering products that are not in stock, by disabling the
add-to-cart button for product with a stock under a certain threshold
and by disabling the quantity-add button when the maximum available
product is chosen.

This works when the product template is correctly configured.

To keep coherence, the module `website_sale_stock_options` should be
installed in order to prevent selling product without checking available
stock in the modal window and in the payment page.